+++ NEWS.md
@@ -5,7 +5,16 @@ The current stable version of Nicotine+ is available at [DOWNLOADS.md](doc/DOWNL
 You can run the latest unstable build of Nicotine+ to test recent changes and bug fixes, see [TESTING.md](doc/TESTING.md).
 
 
-## Version 3.3.1rc1 (Release Candidate 1)
+## Version 3.3.2 (February 25, 2024)
+
+### Corrections
+
+ * Fixed a crash when adding transfers while in ungrouped mode
+ * Fixed an issue where dialog message labels were not read by screen readers
+ * Snap: Fixed screen reader detection for enabling GTK 3
+
+
+## Version 3.3.1 (February 24, 2024)
 
 ### Changes